agents/personal/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
httpx>=0.24.0
fastmcp>=0.2.0
python-dotenv>=0.19.0
supabase>=1.0.3
asyncpg==0.28.0
pyyaml==6.0.1
loguru>=0.5.3
backoff==2.2.1
websockets==11.0.3
jsonschema==4.19.0
cryptography==41.0.3
pytest==7.4.2
requests>=2.26.0
numpy>=1.21.0
scikit-learn>=0.24.2
spacy>=3.7.2
nltk>=3.6.2
textblob>=0.15.3
rake-nltk>=1.0.6
